# docs/swagger.yaml
basePath: /
definitions:
  github_com_deepakgudla_bookvault_internal_dto.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  github_com_deepakgudla_bookvault_internal_dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UserResponse'
    type: object
  github_com_deepakgudla_bookvault_internal_dto.CartItemResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      product:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
      quantity:
        type: integer
      subtotal:
        type: number
      updated_at:
        type: string
    type: object
  github_com_deepakgudla_bookvault_internal_dto.CartResponse:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CartItemResponse'
        type: array
      created_at:
        type: string
      id:
        type: integer
      total:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_deepakgudla_bookvault_internal_dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  github_com_deepakgudla_bookvault_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductImageResponse'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_deepakgudla_bookvault_internal_dto.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  github_com_deepakgudla_bookvault_internal_dto.UpdateCartItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.UpdateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    required:
    - first_name
    - last_name
    type: object
  github_com_deepakgudla_bookvault_internal_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_deepakgudla_bookvault_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:1357
info:
  contact:
    email: deepakgudla35@protonmail.com
    name: Deepak Gudla
    url: https://www.linkedin.com/in/deepakgudla
  description: A modern e-commerce API built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: bookvault-API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AuthResponse'
              type: object
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: User login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the refresh token and log out the user
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: User logout
      tags:
      - Authentication
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AuthResponse'
              type: object
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Refresh access token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AuthResponse'
              type: object
        "400":
          description: Invalid request data or user already exists
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Register a new user
      tags:
      - Authentication
  /api/v1/carts/:
    get:
      description: Retrieve current user's shopping cart with all items
      produces:
      - application/json
      responses:
        "200":
          description: successfully retrieved cart
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: cart not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - Cart
  /api/v1/carts/items:
    post:
      description: users can add products to their cart
      parameters:
      - description: item added to cart
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successfully added products to the cart
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CartResponse'
              type: object
        "400":
          description: invalid request data or insufficient stock
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Add items to cart
      tags:
      - Cart
  /api/v1/carts/items/{id}:
    delete:
      description: users can delete cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted item from the cart
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "400":
          description: invalid cart item
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete items from cart
      tags:
      - Cart
    put:
      description: update the quantity of an item in the user's cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: New quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: cart item updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CartResponse'
              type: object
        "400":
          description: invalid request data or insufficient stock
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update items to cart
      tags:
      - Cart
...
  /api/v1/users/profile:
    get:
      description: Get the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile
      parameters:
      - description: Updated profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UserResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    description: Bearer and JWT token (bearer jwt_token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"